isDot check for Path::clear()

+++ src/Packfire/IO/File/Path.php
@@ -132,10 +132,12 @@ class Path {
                 new \RecursiveDirectoryIterator($this->path),
                 \RecursiveIteratorIterator::CHILD_FIRST);
         foreach($iterator as $path){
-            if($path->isDir()){
-                rmdir((string)$path);
-            }else{
-                unlink((string)$path);
+            if(!$path->isDot()){
+                if($path->isDir()){
+                    rmdir((string)$path);
+                }else{
+                    unlink((string)$path);
+                }
             }
         }
     }
